--- emotional-help/src/main/java/app/emotionalhelp/controller/UserController.java
@@ -1,5 +1,0 @@
-package app.emotionalhelp.controller;
-
-
-public class UserController {
-}
