--- emotional-help/src/main/java/app/emotionalhelp/module/User.java
@@ -1,4 +1,0 @@
-package app.emotionalhelp.module;
-
-public class User {
-}
