--- emotional-help/src/main/java/app/emotionalhelp/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package app.emotionalhelp.repository;
-
-public class UserRepository {
-}
