--- emotional-help/src/main/java/app/emotionalhelp/service/UserService.java
@@ -1,4 +1,0 @@
-package app.emotionalhelp.service;
-
-public interface UserService {
-}
