--- emotional-help/src/main/java/app/emotionalhelp/service/impl/UserServiceImpl.java
@@ -1,4 +1,0 @@
-package app.emotionalhelp.service.impl;
-
-public class UserServiceImpl {
-}
