+++ src/main/java/com/epam/emotionalhelp/controller/UserController.java
@@ -1,0 +1,5 @@
+package com.epam.emotionalhelp.controller;
+
+
+public class UserController {
+}
