+++ src/main/java/com/epam/emotionalhelp/module/User.java
@@ -1,0 +1,4 @@
+package com.epam.emotionalhelp.module;
+
+public class User {
+}
