+++ src/main/java/com/epam/emotionalhelp/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.epam.emotionalhelp.repository;
+
+public class UserRepository {
+}
