+++ src/main/java/com/epam/emotionalhelp/service/UserService.java
@@ -1,0 +1,4 @@
+package com.epam.emotionalhelp.service;
+
+public interface UserService {
+}
