+++ src/main/java/com/epam/emotionalhelp/service/impl/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package com.epam.emotionalhelp.service.impl;
+
+public class UserServiceImpl {
+}
